backend/requirements.txt
flask==1.1.1
flask-restful==0.3.7
flask-cors==3.0.8
flask_sqlalchemy==2.4
psycopg2